requirements.txt
alembic==1.7.6
attrs==21.4.0
Authlib==0.15.5
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
cryptography==36.0.1
dnspython==2.2.0
email-validator==1.1.3
Flask==2.0.3
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
iniconfig==1.1.1
itsdangerous==2.1.0
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.1.0
marshmallow==3.14.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.8
pytest==7.1.2
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
tomli==2.0.1
urllib3==1.26.8
Werkzeug==2.0.3
WTForms==3.0.1
zipp==3.8.0
